LPS-86554 Move to asset.display.contributor package for consistency with the interface
