+++ modules/apps/asset/asset-display-impl/src/main/java/com/liferay/asset/display/internal/asset/display/contributor/AssetEntryAuthorNameAssetDisplayContributorField.java
@@ -12,17 +12,21 @@
  * details.
  */
 
-package com.liferay.asset.display.internal.asset.display.contributor.field;
+package com.liferay.asset.display.internal.asset.display.contributor;
 
 import com.liferay.asset.display.contributor.AssetDisplayContributorField;
 import com.liferay.asset.kernel.model.AssetEntry;
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.service.UserLocalService;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Jürgen Kappler
@@ -31,12 +35,12 @@ import org.osgi.service.component.annotations.Component;
 	property = "model.class.name=com.liferay.asset.kernel.model.AssetEntry",
 	service = AssetDisplayContributorField.class
 )
-public class AssetEntrySummaryAssetDisplayContributorField
+public class AssetEntryAuthorNameAssetDisplayContributorField
 	implements AssetDisplayContributorField<AssetEntry> {
 
 	@Override
 	public String getKey() {
-		return "summary";
+		return "authorName";
 	}
 
 	@Override
@@ -44,7 +48,7 @@ public class AssetEntrySummaryAssetDisplayContributorField
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			locale, getClass());
 
-		return LanguageUtil.get(resourceBundle, "summary");
+		return LanguageUtil.get(resourceBundle, "author-name");
 	}
 
 	@Override
@@ -54,7 +58,18 @@ public class AssetEntrySummaryAssetDisplayContributorField
 
 	@Override
 	public String getValue(AssetEntry assetEntry, Locale locale) {
-		return assetEntry.getSummary(locale);
+		long userId = assetEntry.getUserId();
+
+		User user = _userLocalService.fetchUser(userId);
+
+		if (user != null) {
+			return user.getFullName();
+		}
+
+		return StringPool.BLANK;
 	}
+
+	@Reference
+	private UserLocalService _userLocalService;
 
 }
