+++ modules/apps/asset/asset-display-impl/src/main/java/com/liferay/asset/display/internal/asset/display/contributor/AssetEntryDescriptionAssetDisplayContributorField.java
@@ -12,13 +12,11 @@
  * details.
  */
 
-package com.liferay.asset.display.internal.asset.display.contributor.field;
+package com.liferay.asset.display.internal.asset.display.contributor;
 
 import com.liferay.asset.display.contributor.AssetDisplayContributorField;
 import com.liferay.asset.kernel.model.AssetEntry;
-import com.liferay.asset.kernel.model.AssetTag;
 import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
 import java.util.Locale;
@@ -33,12 +31,12 @@ import org.osgi.service.component.annotations.Component;
 	property = "model.class.name=com.liferay.asset.kernel.model.AssetEntry",
 	service = AssetDisplayContributorField.class
 )
-public class AssetEntryTagsAssetDisplayContributorField
+public class AssetEntryDescriptionAssetDisplayContributorField
 	implements AssetDisplayContributorField<AssetEntry> {
 
 	@Override
 	public String getKey() {
-		return "tagNames";
+		return "description";
 	}
 
 	@Override
@@ -46,7 +44,7 @@ public class AssetEntryTagsAssetDisplayContributorField
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			locale, getClass());
 
-		return LanguageUtil.get(resourceBundle, "tags");
+		return LanguageUtil.get(resourceBundle, "description");
 	}
 
 	@Override
@@ -56,7 +54,7 @@ public class AssetEntryTagsAssetDisplayContributorField
 
 	@Override
 	public String getValue(AssetEntry assetEntry, Locale locale) {
-		return ListUtil.toString(assetEntry.getTags(), AssetTag.NAME_ACCESSOR);
+		return assetEntry.getDescription(locale);
 	}
 
 }
