+++ modules/apps/asset/asset-display-impl/src/main/java/com/liferay/asset/display/internal/asset/display/contributor/AssetEntryTagsAssetDisplayContributorField.java
@@ -12,21 +12,19 @@
  * details.
  */
 
-package com.liferay.asset.display.internal.asset.display.contributor.field;
+package com.liferay.asset.display.internal.asset.display.contributor;
 
 import com.liferay.asset.display.contributor.AssetDisplayContributorField;
 import com.liferay.asset.kernel.model.AssetEntry;
-import com.liferay.petra.string.StringPool;
+import com.liferay.asset.kernel.model.AssetTag;
 import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.service.UserLocalService;
+import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
 
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Jürgen Kappler
@@ -35,12 +33,12 @@ import org.osgi.service.component.annotations.Reference;
 	property = "model.class.name=com.liferay.asset.kernel.model.AssetEntry",
 	service = AssetDisplayContributorField.class
 )
-public class AssetEntryAuthorNameAssetDisplayContributorField
+public class AssetEntryTagsAssetDisplayContributorField
 	implements AssetDisplayContributorField<AssetEntry> {
 
 	@Override
 	public String getKey() {
-		return "authorName";
+		return "tagNames";
 	}
 
 	@Override
@@ -48,7 +46,7 @@ public class AssetEntryAuthorNameAssetDisplayContributorField
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			locale, getClass());
 
-		return LanguageUtil.get(resourceBundle, "author-name");
+		return LanguageUtil.get(resourceBundle, "tags");
 	}
 
 	@Override
@@ -58,18 +56,7 @@ public class AssetEntryAuthorNameAssetDisplayContributorField
 
 	@Override
 	public String getValue(AssetEntry assetEntry, Locale locale) {
-		long userId = assetEntry.getUserId();
-
-		User user = _userLocalService.fetchUser(userId);
-
-		if (user != null) {
-			return user.getFullName();
-		}
-
-		return StringPool.BLANK;
+		return ListUtil.toString(assetEntry.getTags(), AssetTag.NAME_ACCESSOR);
 	}
-
-	@Reference
-	private UserLocalService _userLocalService;
 
 }
