+++ modules/apps/asset/asset-display-impl/src/main/java/com/liferay/asset/display/internal/asset/display/contributor/AssetEntryTitleAssetDisplayContributorField.java
@@ -12,15 +12,12 @@
  * details.
  */
 
-package com.liferay.asset.display.internal.asset.display.contributor.field;
+package com.liferay.asset.display.internal.asset.display.contributor;
 
 import com.liferay.asset.display.contributor.AssetDisplayContributorField;
 import com.liferay.asset.kernel.model.AssetEntry;
 import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.util.FastDateFormatFactoryUtil;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
-
-import java.text.Format;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
@@ -34,12 +31,12 @@ import org.osgi.service.component.annotations.Component;
 	property = "model.class.name=com.liferay.asset.kernel.model.AssetEntry",
 	service = AssetDisplayContributorField.class
 )
-public class AssetEntryPublishDateAssetDisplayContributorField
+public class AssetEntryTitleAssetDisplayContributorField
 	implements AssetDisplayContributorField<AssetEntry> {
 
 	@Override
 	public String getKey() {
-		return "publishDate";
+		return "title";
 	}
 
 	@Override
@@ -47,7 +44,7 @@ public class AssetEntryPublishDateAssetDisplayContributorField
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			locale, getClass());
 
-		return LanguageUtil.get(resourceBundle, "publish-date");
+		return LanguageUtil.get(resourceBundle, "title");
 	}
 
 	@Override
@@ -57,10 +54,7 @@ public class AssetEntryPublishDateAssetDisplayContributorField
 
 	@Override
 	public String getValue(AssetEntry assetEntry, Locale locale) {
-		Format dateFormatDateTime = FastDateFormatFactoryUtil.getDateTime(
-			locale);
-
-		return dateFormatDateTime.format(assetEntry.getPublishDate());
+		return assetEntry.getTitle(locale);
 	}
 
 }
